IA Prototype 2
This newer version of the code goes on to ask the user to input the number of students in a certain class, then it will ask for their names along with grade. Their grades will be asked for after each individual name is inputted and then organized highest to lowest. My next plan is to update the code so it can display the data for each individual student such as GPA and year they are in.